added dual Ethernet/USB support on wiznet6300

+++ wiznet-io/common/pico_cli.h
@@ -36,8 +36,8 @@ static inline void pico_cli_show(const pico_config_t *c, char *out, int outsz)
     if (obs_mirror_enabled(c)) snprintf(obs, sizeof obs, "%d", obs_mirror_pin(c));
     else                       snprintf(obs, sizeof obs, "off");
     int k = snprintf(out, outsz,
-        "name=%s net.mode=%s net.ip=%u.%u.%u.%u dserv=%u.%u.%u.%u:%u obs.pin=%s wifi.ssid=%s wifi.pass=%s wifi.pm=%u ain.en=%u ain.rate=%u ain.gain=%u applied=%u\r\n",
-        dserv_cfg_name(c), dserv_netmode_str(c->net_mode),
+        "name=%s net.mode=%s transport=%s net.ip=%u.%u.%u.%u dserv=%u.%u.%u.%u:%u obs.pin=%s wifi.ssid=%s wifi.pass=%s wifi.pm=%u ain.en=%u ain.rate=%u ain.gain=%u applied=%u\r\n",
+        dserv_cfg_name(c), dserv_netmode_str(c->net_mode), dserv_xport_str(c->transport_mode),
         c->net_ip[0], c->net_ip[1], c->net_ip[2], c->net_ip[3],
         c->dserv_ip[0], c->dserv_ip[1], c->dserv_ip[2], c->dserv_ip[3],
         c->dserv_port, obs,
@@ -157,13 +157,19 @@ static inline cli_action_t pico_cli_exec(pico_config_t *c, const char *line,
         else { snprintf(out, outsz, "ERR net mode dhcp|static\r\n"); return CLI_ERR; }
         c->applied_count++; snprintf(out, outsz, "OK net mode=%s (save+reboot to apply)\r\n", dserv_netmode_str(c->net_mode)); return CLI_OK;
     }
+    if (sscanf(line, "mode %11s", w) == 1) {      /* dual build: transport override (save+reboot) */
+        int m = dserv_xport_val(w);
+        if (m < 0) { snprintf(out, outsz, "ERR mode auto|eth|usb\r\n"); return CLI_ERR; }
+        c->transport_mode = (uint8_t) m; c->applied_count++;
+        snprintf(out, outsz, "OK transport mode=%s (save+reboot to apply)\r\n", dserv_xport_str((uint8_t)m)); return CLI_OK;
+    }
     if (!strcmp(line, "show"))    { pico_cli_show(c, out, outsz); return CLI_OK; }
     if (!strcmp(line, "save"))    { snprintf(out, outsz, "saving...\r\n"); return CLI_SAVE; }
     if (!strcmp(line, "factory")) { snprintf(out, outsz, "factory reset...\r\n"); return CLI_FACTORY; }
     if (!strcmp(line, "reboot"))  { snprintf(out, outsz, "rebooting...\r\n"); return CLI_REBOOT; }
     if (!strcmp(line, "help")) {
         snprintf(out, outsz,
-            "cmds: show | name NAME | net mode dhcp|static | net ip A.B.C.D |\r\n"
+            "cmds: show | name NAME | mode auto|eth|usb | net mode dhcp|static | net ip A.B.C.D |\r\n"
             "      wifi ssid SSID | wifi pass PASS | wifi pm 0|1 | dserv ip A.B.C.D | dserv port N |\r\n"
             "      pin N mode out|in|in_pullup|off | pin N pulse US | pin N debounce MS |\r\n"
             "      pin N active_low 0|1 | obs pin N | obs off |\r\n"
